+++ src/render_static/__init__.py
@@ -17,10 +17,10 @@ r"""
 
 """
 
-VERSION = (3, 4, 0)
+VERSION = (3, 5, 0)
 
 __title__ = "Django Render Static"
 __version__ = ".".join(str(i) for i in VERSION)
 __author__ = "Brian Kohan"
 __license__ = "MIT"
-__copyright__ = "Copyright 2020-2025 Brian Kohan"
+__copyright__ = "Copyright 2020-2026 Brian Kohan"