+++ src/render_static/management/commands/renderstatic.py
@@ -15,8 +15,6 @@ import typing as t
 from pathlib import Path
 from typing import Annotated
 
-from click import Context, Parameter
-from click.shell_completion import CompletionItem
 from django.core.management.base import CommandError
 from django.utils.translation import gettext as _
 from django_typer.completers import chain
@@ -29,34 +27,7 @@ from django_typer.management import TyperCommand
 from typer import Argument, Option
 
 from render_static.engine import StaticTemplateEngine
-
-
-def complete_selector(
-    ctx: Context, param: Parameter, incomplete: str
-) -> t.List[CompletionItem]:
-    """
-    Generate completions for the template selectors.
-    """
-    engine = StaticTemplateEngine()
-    present = ctx.params.get(param.name or "") or []
-    completions = []
-    seen = set()
-    for template in engine.search(
-        incomplete,
-        first_engine=bool(ctx.params.get("first_engine")),
-        first_loader=bool(ctx.params.get("first_loader")),
-    ):
-        tmpl_name = str(template.origin.template_name or "")
-        if tmpl_name and tmpl_name not in present and tmpl_name not in seen:
-            # the slicing is because we need to denormalize the prefix if the
-            # search process normalized the name somehow, because the prefixes
-            # must exactly match whats on the command line for most shell completion
-            # utilities
-            completions.append(
-                CompletionItem(f"{incomplete}{tmpl_name[len(incomplete) :]}")
-            )
-            seen.add(tmpl_name)
-    return completions
+from render_static.management.completers import complete_selector
 
 
 class Command(TyperCommand):